SelectJdbcTemplate, JdbcTemplate 통합
